fix: renamed product module to order
