+++ src/app.module.ts
@@ -4,7 +4,7 @@ import { ConfigModule } from '@nestjs/config';
 import { MysqlModule } from 'src/common/configs/database/mysql.module';
 import { AuthModule } from 'src/modules/auth/auth.module';
 import { GoogleAuthModule } from 'src/modules/google-auth/google-auth.module';
-import { OrderModule } from 'src/modules/product/order.module';
+import { OrderModule } from 'src/modules/order/order.module';
 import { RepositoryModule } from 'src/modules/repository/repository.module';
 import { UserModule } from 'src/modules/user/user.module';
