+++ src/modules/order/order.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 
-import { OrderController } from 'src/modules/product/order.controller';
-import { OrderService } from 'src/modules/product/order.service';
+import { OrderController } from 'src/modules/order/order.controller';
+import { OrderService } from 'src/modules/order/order.service';
 
 @Module({
   imports: [],
