+++ src/modules/user/services/follow.service.ts
@@ -1,4 +1,5 @@
 import { ConflictException, Injectable } from '@nestjs/common';
+
 import { EErrorMessage } from 'src/common/enums/error-message.enum';
 import { FollowRepository } from 'src/modules/repository/services/followe.repository';
 import { UserService } from 'src/modules/user/services/user.service';
